requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=10.0.0
matplotlib>=3.7.0
pypylon>=3.0.0
